requirements.txt
python-telegram-bot
requests
bottle